% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{atac_examples}
\alias{atac_examples}
\title{A small example version of the PBMC ATAC dataset}
\format{
An object of class \code{list} of length 4.
}
\usage{
atac_examples
}
\description{
A small example version of the PBMC ATAC dataset
}
\keyword{datasets}
