% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat-matrices}
\alias{concat-matrices}
\alias{concatByRows}
\alias{concatByCols}
\alias{concatByRows.matrix}
\alias{concatByCols.matrix}
\alias{concatByRows.dgCMatrix}
\alias{concatByCols.dgCMatrix}
\title{Concatenate matrices}
\usage{
concatByRows(x, y, ...)

concatByCols(x, y, ...)

\method{concatByRows}{matrix}(x, y, ...)

\method{concatByCols}{matrix}(x, y, ...)

\method{concatByRows}{dgCMatrix}(x, y, ...)

\method{concatByCols}{dgCMatrix}(x, y, ...)
}
\arguments{
\item{x}{A matrix-like object}

\item{y}{A single or a list of matrix-like object(s) to be merged.}

\item{...}{Arguments passed to other metheds.}
}
\value{
A concatenated matrix-like object.
}
\description{
Methods to concatenate matrix-like objects by rows or columns.
}
\details{
When concatenate matrices by rows, the union of original column names will be
used. Any duplicated row names among matrices will be enforced unique by
adding suffixes with \code{\link{make.unique}}.

When concatenate matrices by columns, use the opposite way.
}
