% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concatByCols-objects}
\alias{concatByCols-objects}
\alias{concatByCols.SummarizedExperiment}
\alias{concatByCols.SingleCellExperiment}
\alias{concatByCols.FlexExperiment}
\alias{concatByCols.SingleCellFlexExperiment}
\title{Concatenate complex matrix-like objects}
\usage{
\method{concatByCols}{SummarizedExperiment}(
  x,
  y,
  assays = NULL,
  label = NULL,
  add.idx = NULL,
  collapse = "_",
  verbose = TRUE,
  ...
)

\method{concatByCols}{SingleCellExperiment}(
  x,
  y,
  assays = NULL,
  reducedDims = NULL,
  altExps = NULL,
  altExps.assays = NULL,
  label = NULL,
  add.idx = NULL,
  collapse = "_",
  verbose = TRUE,
  ...
)

\method{concatByCols}{FlexExperiment}(
  x,
  y,
  assays = NULL,
  assayClasses = NULL,
  rowFlex = "bounded",
  colFlex = "bounded",
  label = NULL,
  add.idx = NULL,
  collapse = "_",
  verbose = TRUE,
  ...
)

\method{concatByCols}{SingleCellFlexExperiment}(
  x,
  y,
  assays = NULL,
  reducedDims = NULL,
  assayClasses = NULL,
  rowFlex = "bounded",
  colFlex = "bounded",
  label = NULL,
  add.idx = NULL,
  collapse = "_",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A matrix-like object}

\item{y}{A single or a list of matrix-like object(s) to be merged.}

\item{assays}{Names of \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays}} to be merged. If set to
\code{NULL}, common assays will be automatically searched from input objects. If
set to \code{FALSE}, no assay will be merged.}

\item{label}{Name of a new column in \code{colData()} to hold the batch
information. Default is \code{NULL} (No batch information added).}

\item{add.idx}{A character vector with \code{length(c(x, y))} elements; appends
the corresponding values to the start of each objects' cell names.}

\item{collapse}{String to collapse \code{add.idx} and cell names of each input
object. Default is \code{"_"}.}

\item{verbose}{Print progress.}

\item{...}{Arguments passed to other metheds.}

\item{altExps}{Names of \code{\link[SingleCellExperiment]{altExps}} to be merged. If set to
\code{NULL}, common alternative experiments will be automatically searched from
input objects. If set to \code{FALSE}, no \code{altExps} will be merged.}

\item{altExps.assays}{Names of \code{assays} to be merged in \code{altExps}.}

\item{assayClasses, rowFlex, colFlex}{Parameters passed to the new merged
\code{\link{FlexExperiment}}.}
}
\value{
A concatenated matrix-like object.
}
\description{
Methods to concatenate matrix-like objects to merge omics data.
}
