% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexExperiment-class.R,
%   R/SingleCellFlexExperiment-class.R
\name{FlexExperiment-subset}
\alias{FlexExperiment-subset}
\alias{[,FlexExperiment,ANY,ANY,ANY-method}
\alias{[[,FlexExperiment,ANY,ANY-method}
\alias{[[<-,FlexExperiment,ANY,ANY-method}
\alias{.DollarNames.FlexExperiment}
\alias{$.FlexExperiment}
\alias{$<-.FlexExperiment}
\alias{[,SingleCellFlexExperiment,ANY,ANY,ANY-method}
\title{Extraction methods for FlexExperiment}
\usage{
\S4method{[}{FlexExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{FlexExperiment,ANY,ANY}(x, i, j, ...)

\S4method{[[}{FlexExperiment,ANY,ANY}(x, i, j, ...) <- value

\method{.DollarNames}{FlexExperiment}(x, pattern = "")

\method{$}{FlexExperiment}(x, name)

\method{$}{FlexExperiment}(x, name) <- value

\S4method{[}{SingleCellFlexExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{\link{FlexExperiment}} object.}

\item{i, j}{Indices specifying elements to extract or replace.}

\item{...}{Arguments passed to other metheds.}

\item{drop}{Whether or not to drop the empty assays when subset \code{x}.}

\item{value}{An object of a class specified in the S4 method signature.}

\item{patten}{A regular expression. Only matching names are returned.}
}
\value{
\itemize{
\item \code{[}: A subset \code{x}.
}

\itemize{
\item \code{[[} or \code{$}: One column of \code{colData(x)}.
}

\itemize{
\item \verb{[[<-} or \verb{$<-}: An updated \code{x} with \code{value} added to one specific
column of \code{colData(x)}.
}
}
\description{
Methods for extraction or subsetting \code{\link{FlexExperiment}} with
\code{[}, \code{[[} or \code{$}.
}
