% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexExperiment-class.R
\docType{class}
\name{FlexExperiment}
\alias{FlexExperiment}
\title{The FlexExperiment class}
\description{
The \code{FlexExperiment} class is a container where matrix-like objects are
stored in a \code{\link{FlexAssays}}. The rows represent features of
interest (e.g. genes, etc...) and columns represent samples (cells).
}
\details{
In \code{FlexExperiment}, both row names and column names are only stored in
\code{assays}.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A \code{\link{FlexAssays}} object containing the layered
matrix-like objects.}

\item{\code{rowData,colData}}{\code{\link[S4Vectors]{DataFrame}} that summarize the metadata
of rows/columns. The row number of \verb{row(col)Data} must match the rows/columns
in \code{assays}.}

\item{\code{int_colData,int_elementMetadata}}{\code{\link[S4Vectors]{DataFrame}} of internal
column/row metadata, with number of rows equal to \code{dim(x)}.}

\item{\code{int_metadata}}{A list of internal metadata}
}}

