% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_accessors.R
\name{internal-helpers}
\alias{internal-helpers}
\alias{getInternalAll}
\alias{setInternalAll}
\alias{getInternalNames}
\alias{setInternalNames}
\alias{getInternalOne}
\alias{setInternalOne}
\title{Helper functions for internal accessors}
\usage{
getInternalAll(x, getfun, key)

setInternalAll(x, value, getfun, setfun, key, prefix = NULL)

getInternalNames(x, getfun, key)

setInternalNames(x, value, getfun, setfun, key, funstr, prefix = NULL)

getInternalOne(x, i, getfun, key, funstr)

setInternalOne(
  x,
  i,
  value,
  getfun,
  setfun,
  key,
  namestr,
  substr,
  funstr,
  prefix = NULL
)
}
\arguments{
\item{x}{An object}

\item{getfun}{A function to fetch the internal \code{\link[S4Vectors]{DataFrame}}.
Typically one of the following:
\itemize{
\item \code{\link[SingleCellExperiment:SCE-internals]{int_colData}}
\item \code{\link[SingleCellExperiment:SCE-internals]{int_elementMetadata}}
}}

\item{key}{Name of the fetched \code{\link[S4Vectors]{DataFrame}}. For example,
'reducedDims' and 'colPairs'.}

\item{value}{An object to set into the internal \code{\link[S4Vectors]{DataFrame}}.}

\item{setfun}{A function to set the updated \code{\link[S4Vectors]{DataFrame}} back
into \code{x}. #' Typically one of the following:
\itemize{
\item \code{\link[SingleCellExperiment:SCE-internals]{int_colData<-}}
\item \code{\link[SingleCellExperiment:SCE-internals]{int_elementMetadata<-}}
}}

\item{funstr}{A string specify the name of wrapper function that calls the
corresponding setter function.}

\item{substr}{A string specify the parameter name of \code{index}. For example,
'type' for \code{\link[SingleCellExperiment]{reducedDim}} or \code{\link[SingleCellExperiment]{colPair}}}

\item{...}{Arguments passed to other metheds.}

\item{name.prefix}{A string specify the prefix for unnamed elements. Default
is 'unnamed'.}

\item{index}{Which element to fetch or set the internal data.}

\item{type}{A string specify the type of \code{index}. For example, 'character' or
'numeric'.}
}
\description{
Functions to access internal fields for any class that has implement methods
of \code{\link[SingleCellExperiment:SCE-internals]{int_colData}} and \code{\link[SingleCellExperiment:SCE-internals]{int_elementMetadata}}.
}
