% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{load-examples}
\alias{load-examples}
\alias{exampleFE}
\alias{exampleSCFE}
\alias{exampleSCE}
\alias{load_example_SCE}
\title{Load the example datasets}
\usage{
exampleFE(sample = NULL, type = c("RNA", "ATAC"))

exampleSCFE(sample = NULL, type = c("RNA", "ATAC"))

exampleSCE(sample = NULL, type = c("RNA", "ATAC"))

load_example_SCE(sample = NULL, type = c("RNA", "ATAC"))
}
\arguments{
\item{sample}{One of the following sample names:
\itemize{
\item "pbmc3k_unsorted"
\item "pbmc3k_sorted"
\item "pbmc10k_unsorted"
\item "pbmc10k_sorted"
}}
}
\description{
Four subsetted versions of 10x Genomics' PBMC multi-omics datasets
}
