% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NearestNeighbors-class.R
\name{NearestNeighbors-methods}
\alias{NearestNeighbors-methods}
\alias{nnidx}
\alias{nnidx,NearestNeighbors-method}
\alias{nndist}
\alias{nndist,NearestNeighbors-method}
\alias{alg.idx}
\alias{alg.info}
\alias{alg.idx.NearestNeighbors}
\alias{alg.info.NearestNeighbors}
\alias{as.matrix,NearestNeighbors-method}
\alias{show,NearestNeighbors-method}
\alias{names,NearestNeighbors-method}
\alias{names<-,NearestNeighbors,character-method}
\alias{names<-,NearestNeighbors,NULL-method}
\alias{dim,NearestNeighbors-method}
\alias{dimnames,NearestNeighbors-method}
\alias{dimnames<-,NearestNeighbors,list-method}
\alias{parallel_slot_names,NearestNeighbors-method}
\title{NearestNeighbors methods}
\usage{
nnidx(x, ...)

\S4method{nnidx}{NearestNeighbors}(x, withDimnames = TRUE, ...)

nndist(x, ...)

\S4method{nndist}{NearestNeighbors}(x, withDimnames = TRUE, ...)

alg.idx(object, ...)

alg.info(object, ...)

\method{alg.idx}{NearestNeighbors}(object, ...)

\method{alg.info}{NearestNeighbors}(object, ...)

\S4method{as.matrix}{NearestNeighbors}(x, use.dist = FALSE, ...)

\S4method{show}{NearestNeighbors}(object)

\S4method{names}{NearestNeighbors}(x)

\S4method{names}{NearestNeighbors,character}(x) <- value

\S4method{names}{NearestNeighbors,NULL}(x) <- value

\S4method{dim}{NearestNeighbors}(x)

\S4method{dimnames}{NearestNeighbors}(x)

\S4method{dimnames}{NearestNeighbors,list}(x) <- value

\S4method{parallel_slot_names}{NearestNeighbors}(x)
}
\arguments{
\item{...}{Arguments passed to other metheds.}

\item{withDimnames}{Whether or not to add the names of observations.}

\item{object, x}{A \code{\link{NearestNeighbors}} object}

\item{use.dist}{Whether or not to use the distances for coercing into matrix.}
}
\value{
\itemize{
\item \code{names}: A character vector specifying names of the observations.
}

\itemize{
\item \verb{names<-}: An updated \code{NearestNeighbors} with new observation names.
}

\itemize{
\item \code{dim}: A two-element vector represent the number of observations and
the number of nearest neighbors for each observation.
}

\itemize{
\item \code{dimnames}: The names of observations. Note that \code{colnames(x)} will
always be \code{NULL}.
}

\itemize{
\item \verb{dimnames<-}: An updated \code{x} with new observation names. Note that
\verb{colnames(x)<-} will be ignored.
}
}
\description{
Methods to manipulate a \code{\link{NearestNeighbors}} object
}
\examples{
# Create a NearestNeighbors
nnidx <- matrix(ceiling(runif(100) * 10), nrow = 20, ncol = 5)
nndist <- matrix(runif(100), nrow = 20, ncol = 5)
nn <- NearestNeighbors(nnidx, nndist)

names(nn) <- letters[1:nrow(nn)]

nnidx(nn)

nndist(nn)

as.matrix(nn)
as.matrix(nn, use.dist = TRUE)

names(nn)

dim(nn)

dimnames(nn)

rownames(nn) <- sample(letters, nrow(nn))

nn[c(1, 3, 4, 5)]
nn[sample(names(nn), 10)]

}
