% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NearestNeighbors-class.R
\docType{class}
\name{NearestNeighbors}
\alias{NearestNeighbors}
\title{The NearestNeighbors class}
\usage{
NearestNeighbors(
  nnidx,
  nndist = NULL,
  alg.idx = NULL,
  alg.info = list(),
  NAMES = NULL
)
}
\arguments{
\item{nnidx}{A \code{\link{matrix}} containing the nearest neighbor indices.}

\item{nndist}{A \code{\link{matrix}} containing the distances between those
observations in \code{nnidx}.}

\item{alg.idx}{A pre-built neighbor searching index. E.g. the \code{annoy} index.}

\item{alg.info}{Any information associated with the algorithm that may be
needed downstream.}

\item{NAMES}{Names of observations for which the nearest neighbors have been
computed.}
}
\description{
The \code{NearestNeighbors} class is designed to store the results of nearest
neighbor searching algorithms.
}
\examples{
# Create a NearestNeighbors
nnidx <- matrix(ceiling(runif(100) * 10), nrow = 20, ncol = 5)
nndist <- matrix(runif(100), nrow = 20, ncol = 5)
nn <- NearestNeighbors(nnidx, nndist)

}
