% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_manipulation.R
\name{nn-coerce}
\alias{nn-coerce}
\alias{nnToGraph}
\alias{nnToPairedHits}
\alias{graphToNN}
\alias{graphToNN.sparseMatrix}
\alias{hitsToNN}
\title{Convert Between K Nearest Neighbors (KNN) and Graph}
\usage{
nnToGraph(
  nn.idx,
  nn.weights = NULL,
  repr = c("C", "T"),
  self.loops = TRUE,
  ...
)

nnToPairedHits(nn.idx, nn.weights = NULL, self.loops = TRUE)

graphToNN(object, ...)

\method{graphToNN}{sparseMatrix}(object, ...)

hitsToNN(object, ...)
}
\arguments{
\item{nn.idx}{A \code{\link{matrix}} for observations x K neighbor indices.}

\item{nn.weights}{\code{NULL} (default) or a \code{\link{matrix}} for distances
corresponding to \code{nn.idx}.}

\item{repr}{Which format of the returned sparse matrix:
\itemize{
\item \code{"C"} for \code{\link[Matrix:CsparseMatrix-class]{CsparseMatrix}}
\item \code{"T"} for \code{\link[Matrix:TsparseMatrix-class]{TsparseMatrix}}
}}

\item{self.loops}{Whether to allow self-loops in the output sparse matrix.}

\item{...}{Arguments passed to other metheds.}

\item{object}{An object to be converted into KNN indices and weights.}
}
\value{
\itemize{
\item \code{nnToGraph}: Returns a sparse matrix indicating the KNN graph. Each
column must contain the same number (k) of non-zero values.
}

\itemize{
\item \code{nnToPairedHits}: Returns a \code{\link{PairedHits}} indicating the KNN
graph. Each left node should contain k right nodes.
}

\itemize{
\item \code{graphToNN} and \code{hitsToNN}: Returns a list containing two
\code{\link{matrix}}. \code{idx} indicates the KNN indices and \code{dist} indicates
the weights (distances) between observations and their nearest neighbors.
}
}
\description{
Convert K Nearest Neighbors (KNN) from / to sparse graph.
}
