% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairedHits-class.R
\name{PairedHits-coerce}
\alias{PairedHits-coerce}
\alias{hitsToMat}
\alias{hitsToMat.Hits}
\alias{hitsToMat.PairedHits}
\alias{matToHits}
\alias{matToHits.default}
\alias{matToPairedHits}
\alias{matToPairedHits.default}
\title{Helpers for Coercing PairedHits}
\usage{
hitsToMat(x, ...)

\method{hitsToMat}{Hits}(x, ...)

\method{hitsToMat}{PairedHits}(x, ...)

matToHits(x, ...)

\method{matToHits}{default}(x, self = TRUE, ...)

matToPairedHits(x, ...)

\method{matToPairedHits}{default}(x, ...)
}
\arguments{
\item{x}{An object to be coerced.}

\item{...}{Arguments passed to other metheds.}
}
\value{
\itemize{
\item \code{hitsToMat}: Returns a sparse matrix where row indices are \code{from(x)}
and column indices are \code{to(x)}.
}

\itemize{
\item \code{matToHits}: Returns a \code{\link[NA]{Hits}} object where left nodes
indicate row indices and right nodes indicate column indices.
}

\itemize{
\item \code{matToPairedHits}: Returns a \code{\link{PairedHits}} object. The input
\code{x} must contain identical row names and column names.
}
}
\description{
Helper functions for coercing between \code{\link{PairedHits}} object and
sparse matrix.
}
