% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairedHits-class.R
\name{PairedHits-methods}
\alias{PairedHits-methods}
\alias{showPairedHits}
\alias{show,PairedHits-method}
\alias{length,PairedHits-method}
\alias{names,PairedHits-method}
\alias{names<-,PairedHits,character-method}
\alias{names<-,PairedHits,NULL-method}
\alias{nLnode,PairedHits-method}
\alias{nRnode,PairedHits-method}
\alias{nnode,PairedHits-method}
\alias{[,PairedHits,ANY,ANY,ANY-method}
\title{Methods for PairedHits}
\usage{
showPairedHits(object)

\S4method{show}{PairedHits}(object)

\S4method{length}{PairedHits}(x)

\S4method{names}{PairedHits}(x)

\S4method{names}{PairedHits,character}(x) <- value

\S4method{names}{PairedHits,NULL}(x) <- value

\S4method{nLnode}{PairedHits}(x)

\S4method{nRnode}{PairedHits}(x)

\S4method{nnode}{PairedHits}(x)

\S4method{[}{PairedHits,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x, object}{A \code{\link{FlexAssays}} object}

\item{value}{An object of a class specified in the S4 method signature. Typically a character vector for node names.}

\item{i, j}{Indices specifying elements to extract. \code{j} will be ignored, since
subsetting a \code{\link{PairedHits}} object with \code{i} will simultaneously
work on left and right nodes of the underlying \code{\link[S4Vectors:Hits-class]{SelfHits}}.}

\item{...}{Arguments passed to other metheds.}

\item{drop}{Ignored by subsetting. This is only used for matching the S4
generic function.}
}
\value{
\itemize{
\item \code{length}: An integer scalar indicating the number of nodes.
}

\itemize{
\item \code{names}: A character vector specifying the node names.
}

\itemize{
\item \verb{names<-}: Returns \code{x} with updated node names.
}

\itemize{
\item \code{[}: A subset \code{\link{PairedHits}} object.
}
}
\description{
A collection of accessor methods for retrieving or modifying basic
information from a \code{\link{PairedHits}} object.
}
