% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairedHits-class.R
\docType{class}
\name{PairedHits}
\alias{PairedHits}
\title{The PairedHits Class}
\usage{
PairedHits(hits, NAMES = NULL)
}
\arguments{
\item{hits}{A \code{\link[S4Vectors:Hits-class]{SelfHits}} object.}

\item{NAMES}{A character vector or \code{NULL}, containing node names for \code{hits}.
The number of elements must be equal to \code{nnode(hits)}. Duplicated node names
are not allowed.}
}
\description{
A class for storing paired hit objects. This structure wraps a
\code{\link[S4Vectors:Hits-class]{SelfHits}} object and an optional character vector for node
names.
}
\section{Slots}{

\describe{
\item{\code{hits}}{A \code{\link[S4Vectors:Hits-class]{SelfHits}} object.}

\item{\code{NAMES}}{A character vector or \code{NULL}, containing node names for \code{hits}.}
}}

\seealso{
\code{\link[S4Vectors:Hits-class]{SelfHits}}
}
