% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{pbmc_examples}
\alias{pbmc_examples}
\title{A small example version of the PBMC dataset}
\format{
An object of class \code{list} of length 4.
}
\usage{
pbmc_examples
}
\description{
A small example version of the PBMC dataset
}
\keyword{datasets}
