% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellFlexExperiment-class.R
\name{SingleCellFlexExperiment-assays}
\alias{SingleCellFlexExperiment-assays}
\alias{counts,SingleCellFlexExperiment-method}
\alias{counts<-,SingleCellFlexExperiment-method}
\alias{logcounts,SingleCellFlexExperiment-method}
\alias{logcounts<-,SingleCellFlexExperiment-method}
\alias{normcounts,SingleCellFlexExperiment-method}
\alias{normcounts<-,SingleCellFlexExperiment-method}
\alias{cpm,SingleCellFlexExperiment-method}
\alias{cpm<-,SingleCellFlexExperiment-method}
\alias{tpm,SingleCellFlexExperiment-method}
\alias{tpm<-,SingleCellFlexExperiment-method}
\title{Named assay getters and setters}
\usage{
\S4method{counts}{SingleCellFlexExperiment}(object, ...)

\S4method{counts}{SingleCellFlexExperiment}(object, ...) <- value

\S4method{logcounts}{SingleCellFlexExperiment}(object, ...)

\S4method{logcounts}{SingleCellFlexExperiment}(object, ...) <- value

\S4method{normcounts}{SingleCellFlexExperiment}(object, ...)

\S4method{normcounts}{SingleCellFlexExperiment}(object, ...) <- value

\S4method{cpm}{SingleCellFlexExperiment}(object, ...)

\S4method{cpm}{SingleCellFlexExperiment}(object, ...) <- value

\S4method{tpm}{SingleCellFlexExperiment}(object, ...)

\S4method{tpm}{SingleCellFlexExperiment}(object, ...) <- value
}
\description{
These are convenient methods for getting or setting \code{assay(x, i = X, ...)}
where \code{x} is a \code{\link{SingleCellFlexExperiment}} object.
}
