% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellFlexExperiment-class.R
\name{SingleCellFlexExperiment-methods}
\alias{SingleCellFlexExperiment-methods}
\alias{show,SingleCellFlexExperiment-method}
\alias{reducedDims,SingleCellFlexExperiment-method}
\alias{reducedDims<-,SingleCellFlexExperiment-method}
\alias{reducedDimNames,SingleCellFlexExperiment-method}
\alias{reducedDimNames<-,SingleCellFlexExperiment,character-method}
\alias{reducedDim,SingleCellFlexExperiment,missing-method}
\alias{reducedDim,SingleCellFlexExperiment,integer-method}
\alias{reducedDim,SingleCellFlexExperiment,numeric-method}
\alias{reducedDim,SingleCellFlexExperiment,character-method}
\alias{reducedDim<-,SingleCellFlexExperiment,missing-method}
\alias{reducedDim<-,SingleCellFlexExperiment,integer-method}
\alias{reducedDim<-,SingleCellFlexExperiment,numeric-method}
\alias{reducedDim<-,SingleCellFlexExperiment,character-method}
\alias{colPairs,SingleCellFlexExperiment-method}
\alias{colPairs<-,SingleCellFlexExperiment-method}
\alias{colPairNames,SingleCellFlexExperiment-method}
\alias{colPairNames<-,SingleCellFlexExperiment,character-method}
\alias{colPair,SingleCellFlexExperiment,missing-method}
\alias{colPair,SingleCellFlexExperiment,integer-method}
\alias{colPair,SingleCellFlexExperiment,numeric-method}
\alias{colPair,SingleCellFlexExperiment,character-method}
\alias{colPair<-,SingleCellFlexExperiment,missing-method}
\alias{colPair<-,SingleCellFlexExperiment,integer-method}
\alias{colPair<-,SingleCellFlexExperiment,numeric-method}
\alias{colPair<-,SingleCellFlexExperiment,character-method}
\alias{rowPairs,SingleCellFlexExperiment-method}
\alias{rowPairs<-,SingleCellFlexExperiment-method}
\alias{rowPairNames,SingleCellFlexExperiment-method}
\alias{rowPairNames<-,SingleCellFlexExperiment,character-method}
\alias{rowPair,SingleCellFlexExperiment,missing-method}
\alias{rowPair,SingleCellFlexExperiment,integer-method}
\alias{rowPair,SingleCellFlexExperiment,numeric-method}
\alias{rowPair,SingleCellFlexExperiment,character-method}
\alias{rowPair<-,SingleCellFlexExperiment,missing-method}
\alias{rowPair<-,SingleCellFlexExperiment,integer-method}
\alias{rowPair<-,SingleCellFlexExperiment,numeric-method}
\alias{rowPair<-,SingleCellFlexExperiment,character-method}
\title{Methods for SingleCellFlexExperiment}
\usage{
\S4method{show}{SingleCellFlexExperiment}(object)

\S4method{reducedDims}{SingleCellFlexExperiment}(x, withDimnames = TRUE, ...)

\S4method{reducedDims}{SingleCellFlexExperiment}(x, withDimnames = TRUE, ...) <- value

\S4method{reducedDimNames}{SingleCellFlexExperiment}(x)

\S4method{reducedDimNames}{SingleCellFlexExperiment,character}(x) <- value

\S4method{reducedDim}{SingleCellFlexExperiment,missing}(x, type, withDimnames = TRUE, ...)

\S4method{reducedDim}{SingleCellFlexExperiment,integer}(x, type, withDimnames = TRUE, ...)

\S4method{reducedDim}{SingleCellFlexExperiment,numeric}(x, type, withDimnames = TRUE, ...)

\S4method{reducedDim}{SingleCellFlexExperiment,character}(x, type, withDimnames = TRUE, ...)

\S4method{reducedDim}{SingleCellFlexExperiment,missing}(x, type, withDimnames = TRUE, ...) <- value

\S4method{reducedDim}{SingleCellFlexExperiment,integer}(x, type, withDimnames = TRUE, ...) <- value

\S4method{reducedDim}{SingleCellFlexExperiment,numeric}(x, type, withDimnames = TRUE, ...) <- value

\S4method{reducedDim}{SingleCellFlexExperiment,character}(x, type, withDimnames = TRUE, ...) <- value

\S4method{colPairs}{SingleCellFlexExperiment}(x, asSparse = FALSE, ...)

\S4method{colPairs}{SingleCellFlexExperiment}(x) <- value

\S4method{colPairNames}{SingleCellFlexExperiment}(x)

\S4method{colPairNames}{SingleCellFlexExperiment,character}(x) <- value

\S4method{colPair}{SingleCellFlexExperiment,missing}(x, type, asSparse = FALSE, ...)

\S4method{colPair}{SingleCellFlexExperiment,integer}(x, type, asSparse = FALSE, ...)

\S4method{colPair}{SingleCellFlexExperiment,numeric}(x, type, asSparse = FALSE, ...)

\S4method{colPair}{SingleCellFlexExperiment,character}(x, type, asSparse = FALSE, ...)

\S4method{colPair}{SingleCellFlexExperiment,missing}(x, type, ...) <- value

\S4method{colPair}{SingleCellFlexExperiment,integer}(x, type, ...) <- value

\S4method{colPair}{SingleCellFlexExperiment,numeric}(x, type, ...) <- value

\S4method{colPair}{SingleCellFlexExperiment,character}(x, type, ...) <- value

\S4method{rowPairs}{SingleCellFlexExperiment}(x, asSparse = FALSE, ...)

\S4method{rowPairs}{SingleCellFlexExperiment}(x) <- value

\S4method{rowPairNames}{SingleCellFlexExperiment}(x)

\S4method{rowPairNames}{SingleCellFlexExperiment,character}(x) <- value

\S4method{rowPair}{SingleCellFlexExperiment,missing}(x, type, asSparse = FALSE, ...)

\S4method{rowPair}{SingleCellFlexExperiment,integer}(x, type, asSparse = FALSE, ...)

\S4method{rowPair}{SingleCellFlexExperiment,numeric}(x, type, asSparse = FALSE, ...)

\S4method{rowPair}{SingleCellFlexExperiment,character}(x, type, asSparse = FALSE, ...)

\S4method{rowPair}{SingleCellFlexExperiment,missing}(x, type, ...) <- value

\S4method{rowPair}{SingleCellFlexExperiment,integer}(x, type, ...) <- value

\S4method{rowPair}{SingleCellFlexExperiment,numeric}(x, type, ...) <- value

\S4method{rowPair}{SingleCellFlexExperiment,character}(x, type, ...) <- value
}
\arguments{
\item{x, object}{A \code{\link{SingleCellFlexExperiment}} object.}

\item{...}{Arguments passed to other metheds.}
}
\description{
Accessor methods specific for \code{\link{SingleCellFlexExperiment}}.
}
\seealso{
S4 generics defined in \pkg{SingleCellExperiment}:
\itemize{
\item \code{\link[SingleCellExperiment]{reducedDims}}
\item \code{\link[SingleCellExperiment]{colPairs}}
\item \code{\link[SingleCellExperiment]{rowPairs}}
}
S4 generics defined in \pkg{OmicsArtifacts}:
\itemize{
\item \code{\link{colNeighbors}}
}
}
