% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellFlexExperiment-class.R
\docType{class}
\name{SingleCellFlexExperiment}
\alias{SingleCellFlexExperiment}
\title{The SingleCellFlexExperiment}
\usage{
SingleCellFlexExperiment(
  ...,
  reducedDims = list(),
  rowPairs = list(),
  colPairs = list()
)
}
\arguments{
\item{reducedDims}{A list of any number of matrix-like objects representing
dimensionality reduction results, each of which should have the same number
of rows as the output \code{SingleCellFlexExperiment}.}

\item{rowPairs}{A list of any number of \code{\link[Matrix:CsparseMatrix-class]{CsparseMatrix}}
representing the pairing data of rows, each of which should inherit from
\code{dMatrix} and contain the same numbers of rows and columns, matching rows of
the output \code{SingleCellFlexExperiment}.}

\item{colPairs}{A list of any number of \code{\link[Matrix:CsparseMatrix-class]{CsparseMatrix}}
representing the pairing data of columns, each of which should inherit from
\code{dMatrix} and contain the same numbers of rows and columns, matching columns
of the output \code{SingleCellFlexExperiment}.}

\item{assays}{A \code{\link{list}} or \code{\link[NA]{SimpleList}} of
matrix-like objects. Also supports a pre-built \code{\link{LayeredAssays}}.
Names of rows and columns must present in those matrix-like objects.}

\item{rowData}{A \code{\link[S4Vectors]{DataFrame}} object summarizing the rows of
\code{assays}. Row names must present to match each row to corresponding rows in
\code{assays}. If \code{rowData} is not provided, the output row names will use the
union of row names in \code{assays}.}

\item{rowRanges}{A \code{\link[GenomicRanges:GRanges-class]{GRanges}} or \code{\link[GenomicRanges:GRangesList-class]{GRangesList}}
object describing the ranges of interest. The length of \code{rowRanges} must be
equal to the row number of \code{assays}.}

\item{colData}{A \code{\link[S4Vectors]{DataFrame}} object summarizing the columns of
\code{assays}. Row names must present to match each row to corresponding columns
in \code{assays}. If \code{colData} is not provided, the output column names will use
the union of column names in \code{assays}.}

\item{colNeighbors}{A list of any number of \code{\link{NearestNeighbors}}
objects, each of which should have the same number of observations as the
column number of the output \code{SingleCellFlexExperiment}.}

\item{metadata}{An optional \code{\link{list}} of arbitrary content
describing the overall experiment.}
}
\description{
The \code{SingleCellFlexExperiment} class is designed to represent single-cell
data. It inherits from \code{\link{LayeredExperiment}} class and is used in
the same manner. By default, this class supplies storage of dimensionality
reduction results (\code{\link{reducedDims}}), row or column pairing data
(\code{\link{rowPairs}} and \code{\link{colPairs}}) and column nearest
neighbor (NN) results (\code{\link{colNeighbors}}).
}
