% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairedHits-class.R
\name{validPairedHits}
\alias{validPairedHits}
\title{Validation for PairedHits}
\usage{
validPairedHits(x, immediate. = TRUE)
}
\arguments{
\item{x}{A \code{\link{PairedHits}} to check.}

\item{immediate.}{Logical indicating if the errors should be raised
immediately. Set \code{immediate. = FALSE} can be useful for context of
\code{\link[NA]{validObject}}.}
}
\value{
Returns an invisible \code{NULL} if \code{x} is a valid \code{\link{PairedHits}}. If
\code{immediate. = TRUE}, this function will raise an error when \code{x} is invalid.
Otherwise, it will return a character vector that collects all errors.
}
\description{
Check whether a \code{\link{PairedHits}} is valid.
}
\seealso{
\code{\link[=getErrors]{getErrors()}}
}
